Частичная реализация консольного ввода. Попытка собрать main(). Данная реализация не работает, но компилируется

+++ Test1/src/main/java/ru/tests/techtracker/Tracker.java
@@ -113,6 +113,6 @@ public class Tracker {
      * @return Уникальный ключ.
      */
     private String generateId() {
-        return String.valueOf(System.currentTimeMillis() + RN.nextInt());
+        return String.valueOf(System.currentTimeMillis() + RN.nextLong());
     }
 }
